Introduced `getApprover(...)` authorizer interface.
This patch introduces a breaking change in the Authorizer interface:
`getObjectApprover(...)` method that returnes ObjectApprover
which should not be stored for a long time is replaced with
`getApprover(...)` method that returns ObjectApprover
that must be kept valid (by authorizer implementation) throughout its
whole lifetime.

This unblocks way to synchronous (without dispatch to another actor)
authorization in cases where principal is known to be long-lived;
examples are the scheduler API (see MESOS-10056) and v1 operator API
events (see MESOS-10057).

The local authorizer is modified accordingly.

NOTE: This patch breaks compatibility with custom authorizers which
do not implement this method!

Review: https://reviews.apache.org/r/72089